news_batcher/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
apscheduler==3.10.4
google-generativeai==0.3.2
sqlalchemy==2.0.25
pymysql==1.1.0
feedparser==6.0.11
python-dotenv==1.0.0
requests==2.31.0
beautifulsoup4==4.12.3
